Make use of Agency in Departures lookups
Agency already existed as a concept, but was not being used to scope
queries or to figure out the Realtime API to call, etc. Now the code
implicitly uses the Agency from the stop_id passed and limits all of the
relationships to those of the same agency.

Kind of a big commit because it touched a lot of different areas
including how we build our test data.

Signed-off-by: Geoff Lane <geoff@zorched.net>