Use proper start_time for determining offsets
Timezones are hard. You need to understand that Rails converts times to UTC going in
and from UTC to local time coming out. In the DB everything is UTC and has to be
compared that way. BUT... intervals are in local time, so you need to convert those
UTC times to local times.

Signed-off-by: Geoff Lane <geoff@zorched.net>