Import improvements
Do the whole thing in a transaction so that we don't have downtown when things are deleted.
Remove old data before importing to ensure that we outdated stops, etc no longer show up.
Remove find_or_initialize blocks because those are only run when initialize is done and updates
weren't being run in the update case.

Signed-off-by: Geoff Lane <geoff@zorched.net>