Departures end point accepts an end_time_offset
This allows consumers of the API to specify how many hours in the future
to look for stop times. We prefer using offset instead of allowing the 
client to pass in raw end time values so that the server remains the 
source of truth for time. There are cases where the client may have the 
wrong time information and we still want to provide accurate data to 
them.